requirements.txt
Flask==1.1.1
redis==3.5.2
PyMySQL==0.9.3
Flask-WTF==0.14.3
Flask-Login==0.5.0
docker-py==1.10.6
easygui==0.98.1